Moving more stuff into shared libary.
